feat: refactor extract zip utility
