+++ src/types/components/index.ts
@@ -5,9 +5,6 @@ export type {
 	TemplateCardType,
 	DashboardFooterTemplateCardType,
 	RemoteTemplateCategory,
-	RemoteTemplateMetadata,
-	RemoteTemplateCardWithFiles,
-	ProcessedRemoteCategory,
 } from "@src/types/components/projectTemplates.type";
 export type { TopbarType } from "@src/types/components/topbar.type";
 export type { ButtonType } from "@type/components/button.type";
