+++ src/types/components/projectTemplates.type.ts
@@ -1,5 +1,7 @@
 import { ComponentType, SVGProps } from "react";
 
+import { RemoteTemplateMetadata } from "@src/interfaces/store";
+
 export type TemplateCardType = {
 	assetDirectory: string;
 	description: string;
@@ -29,31 +31,3 @@ type Integration = {
 	icon: ComponentType<SVGProps<SVGSVGElement>>;
 	label: string;
 };
-
-export interface TemplateFile {
-	id: string;
-	templateId: string;
-	path: string;
-	content: string;
-}
-
-export interface RemoteTemplateMetadata {
-	assetDirectory: string;
-	title: string;
-	description: string;
-	integrations: string[];
-	filesIndex: string[];
-}
-
-export interface RemoteTemplateCardWithFiles {
-	assetDirectory: string;
-	title: string;
-	description: string;
-	integrations: string[];
-	files: Record<string, string>;
-}
-
-export interface ProcessedRemoteCategory {
-	name: string;
-	cards: RemoteTemplateCardWithFiles[];
-}
